backend/requirement.txt
asgiref==3.8.1
Django==5.1.7
django-cors-headers==4.7.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
packaging==24.2
pdf2image==1.17.0
pillow==11.1.0
psycopg2-binary==2.9.10
PyJWT==2.9.0
PyPDF2==3.0.1
pytesseract==0.3.13
rest-framework-simplejwt==0.0.2
sqlparse==0.5.3
